+++ src/main/java/vendingmachine/utill/ErrorMsgConst.java
@@ -1,0 +1,10 @@
+package vendingmachine.utill;
+
+public class ErrorMsgConst {
+	public static final String ERROR_MSG = "[ERROR] ";
+	public static final String INT_TYPE_ERROR_MSG = "금액은 숫자여야 합니다.";
+	public static final String NEGATIVE_INT_ERROR_MSG = "금액은 0원 이상이어야 합니다.";
+	public static final String NOT_DIVISIBLE_BY_10_ERROR_MSG = "금액은 10원 단위로 나누어 떨어져야 합니다.";
+	public static final String ITEM_INPUT_ERROR_MSG = "상품 정보의 입력 형식이 올바르지 않습니다.";
+	public static final String ITEM_DUP_ERROR_MSG = "상품명은 중복되지 않아야 합니다.";
+}
