+++ src/main/java/vendingmachine/view/InputFromAdmin.java
@@ -4,20 +4,22 @@ import java.util.Arrays;
 import java.util.List;
 
 import camp.nextstep.edu.missionutils.Console;
-import vendingmachine.utill.InputValidator;
+import vendingmachine.utill.InputItemValidator;
+import vendingmachine.utill.InputMoneyValidator;
 
 public class InputFromAdmin {
 	private static final String MACHINE_MONEY_INPUT_MSG = "자판기가 보유하고 있는 금액을 입력해 주세요.";
 	private static final String ITEM_INPUT_MSG = "상품명과 가격, 수량을 입력해 주세요.";
 	private static final String DELIMITER = ";";
 
-	private InputValidator inputValidator = new InputValidator();
+	private InputMoneyValidator inputMoneyValidator = new InputMoneyValidator();
+	private InputItemValidator inputItemValidator = new InputItemValidator();
 
 	public int inputMachineMoney() {
 		System.out.println(MACHINE_MONEY_INPUT_MSG);
 		String inputAmount = Console.readLine();
 
-		int amount = inputValidator.validMachineMoney(inputAmount);
+		int amount = inputMoneyValidator.validMachineMoney(inputAmount);
 
 		if (amount >= 0) {
 			return amount;
@@ -30,7 +32,7 @@ public class InputFromAdmin {
 		String items = Console.readLine();
 		List<String> itemList = splitInputItems(items);
 
-		if (inputValidator.validItemList(itemList)) {
+		if (inputItemValidator.validItemList(itemList)) {
 			return itemList;
 		}
 		return inputItems();
